index.ts -> test.tsに、それに応じてpackage.jsonも変更
